requirements.txt
Django==4.2.3
djangorestframework==3.14.0
crispy-bootstrap5==0.7
django-crispy-forms==2.0
celery==5.3.1
python-decouple==3.8
psycopg2==2.9.6
Pillow==10.0.0
whitenoise==6.5.0
gunicorn==21.2.0
